[on-chain config] organize on-chain config into a dir
