+++ config/config-builder/src/validator_config.rs
@@ -11,7 +11,7 @@ use libra_config::{
     generator::{self, ValidatorSwarm},
 };
 use libra_crypto::{ed25519::Ed25519PrivateKey, PrivateKey, Uniform};
-use libra_types::{discovery_set::DiscoverySet, validator_set::ValidatorSet};
+use libra_types::{discovery_set::DiscoverySet, on_chain_config::ValidatorSet};
 use parity_multiaddr::Multiaddr;
 use rand::{rngs::StdRng, Rng, SeedableRng};
 use std::{collections::HashMap, net::SocketAddr};
