+++ language/functional-tests/src/config/global.rs
@@ -11,7 +11,7 @@ use language_e2e_tests::{
 };
 use libra_config::generator;
 use libra_crypto::PrivateKey;
-use libra_types::validator_set::ValidatorSet;
+use libra_types::on_chain_config::ValidatorSet;
 use std::{
     collections::{btree_map, BTreeMap},
     str::FromStr,
