+++ types/src/on_chain_config/mod.rs
@@ -6,14 +6,21 @@ use crate::{
     account_config::{association_address, CORE_CODE_ADDRESS},
     event::{EventHandle, EventKey},
     language_storage::{StructTag, TypeTag},
-    move_resource::MoveResource,
-    transaction::SCRIPT_HASH_LENGTH,
 };
 use anyhow::{format_err, Result};
-use libra_crypto::HashValue;
 use move_core_types::identifier::Identifier;
 use serde::{de::DeserializeOwned, Deserialize, Serialize};
 use std::{collections::HashMap, sync::Arc};
+
+mod libra_version;
+mod validator_set;
+mod vm_publishing_option;
+
+pub use self::{
+    libra_version::LibraVersion, validator_set::ValidatorSet,
+    vm_publishing_option::VMPublishingOption,
+};
+use crate::move_resource::MoveResource;
 
 /// To register an on-chain config in Rust:
 /// 1. Implement the `OnChainConfig` trait for the Rust representation of the config
@@ -132,64 +139,4 @@ impl ConfigurationResource {
 impl MoveResource for ConfigurationResource {
     const MODULE_NAME: &'static str = "LibraConfig";
     const STRUCT_NAME: &'static str = "Configuration";
-}
-
-/// Defines and holds the publishing policies for the VM. There are three possible configurations:
-/// 1. No module publishing, only whitelisted scripts are allowed.
-/// 2. No module publishing, custom scripts are allowed.
-/// 3. Both module publishing and custom scripts are allowed.
-/// We represent these as an enum instead of a struct since whitelisting and module/script
-/// publishing are mutually exclusive options.
-#[derive(Clone, Debug, Deserialize, PartialEq, Serialize)]
-pub enum VMPublishingOption {
-    /// Only allow scripts on a whitelist to be run
-    Locked(Vec<[u8; SCRIPT_HASH_LENGTH]>),
-    /// Allow custom scripts, but _not_ custom module publishing
-    CustomScripts,
-    /// Allow both custom scripts and custom module publishing
-    Open,
-}
-
-impl OnChainConfig for VMPublishingOption {
-    const IDENTIFIER: &'static str = "ScriptWhitelist";
-
-    fn deserialize_into_config(bytes: &[u8]) -> Result<Self> {
-        lcs::from_bytes::<Vec<u8>>(&bytes)
-            .map_err(|e| {
-                format_err!(
-                    "Failed first round of deserialization for VMPublishingOption: {}",
-                    e
-                )
-            })
-            .and_then(|bytes| Self::deserialize_default_impl(&bytes))
-    }
-}
-
-impl VMPublishingOption {
-    pub fn is_open(&self) -> bool {
-        match self {
-            VMPublishingOption::Open => true,
-            _ => false,
-        }
-    }
-
-    pub fn is_allowed_script(&self, program: &[u8]) -> bool {
-        match self {
-            VMPublishingOption::Open | VMPublishingOption::CustomScripts => true,
-            VMPublishingOption::Locked(whitelist) => {
-                let hash_value = HashValue::from_sha3_256(program);
-                whitelist.contains(hash_value.as_ref())
-            }
-        }
-    }
-}
-
-/// Defines the version of Libra Validator software.
-#[derive(Clone, Debug, Deserialize, PartialEq, Serialize)]
-pub struct LibraVersion {
-    pub major: u64,
-}
-
-impl OnChainConfig for LibraVersion {
-    const IDENTIFIER: &'static str = "LibraVersion";
 }
