+++ types/src/unit_tests/validator_set_test.rs
@@ -1,7 +1,7 @@
 // Copyright (c) The Libra Core Contributors
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::validator_set::ValidatorSet;
+use crate::on_chain_config::ValidatorSet;
 use lcs::test_helpers::assert_canonical_encode_decode;
 use libra_prost_ext::test_helpers::assert_protobuf_encode_decode;
 use proptest::prelude::*;
